Tmp structure
